move user context and roles from net/constants
