+++ pkg/net/dtos/jwt_payload.go
@@ -5,13 +5,13 @@ import (
 	"time"
 
 	"github.com/BEOpenSourceCollabs/EventManagementCore/pkg/logger"
-	"github.com/BEOpenSourceCollabs/EventManagementCore/pkg/net/constants"
+	"github.com/BEOpenSourceCollabs/EventManagementCore/pkg/types"
 	"github.com/golang-jwt/jwt/v5"
 )
 
 type JwtPayload struct {
 	Id   string
-	Role constants.Role
+	Role types.Role
 }
 
 // Sign generates a signed JWT token for given payload using the provided secret.
@@ -46,7 +46,7 @@ func (jwtPayload *JwtPayload) ParseSignedToken(tokenString, secret string) error
 	if claims, ok := token.Claims.(jwt.MapClaims); ok && token.Valid {
 		(*jwtPayload) = JwtPayload{
 			Id:   claims["sub"].(string),
-			Role: constants.Role(claims["role"].(string)),
+			Role: types.Role(claims["role"].(string)),
 		}
 	}
