+++ pkg/net/dtos/login.go
@@ -1,6 +1,7 @@
 package dtos
 
 type Login struct {
+	DTO
 	Email    string `json:"email"`
 	Password string `json:"password"`
 }
