+++ pkg/net/routes/route_helpers.go
@@ -6,9 +6,10 @@ import (
 	"net/http"
 
 	"github.com/BEOpenSourceCollabs/EventManagementCore/pkg/models"
-	"github.com/BEOpenSourceCollabs/EventManagementCore/pkg/net/constants"
 	"github.com/BEOpenSourceCollabs/EventManagementCore/pkg/net/dtos"
 	"github.com/BEOpenSourceCollabs/EventManagementCore/pkg/repository"
+	"github.com/BEOpenSourceCollabs/EventManagementCore/pkg/service"
+	"github.com/BEOpenSourceCollabs/EventManagementCore/pkg/types"
 )
 
 type UserContextHelpers struct {
@@ -17,7 +18,7 @@ type UserContextHelpers struct {
 
 // LoadUserFromContext helper that attempts to read the http.Request's user context key or returns an error if it was not found.
 func (h UserContextHelpers) LoadUserFromContext(r *http.Request) (*models.UserModel, error) {
-	userContext, ok := r.Context().Value(constants.USER_CONTEXT_KEY).(*dtos.JwtPayload)
+	userContext, ok := r.Context().Value(service.USER_CONTEXT_KEY).(*dtos.JwtPayload)
 	if !ok {
 		return nil, ErrMissingUserContext
 	}
@@ -30,8 +31,8 @@ func (h UserContextHelpers) LoadUserFromContext(r *http.Request) (*models.UserMo
 	return user, nil
 }
 
-func (h UserContextHelpers) LoadUserFromContextWithRole(r *http.Request, role constants.Role) (*models.UserModel, error) {
-	userContext, ok := r.Context().Value(constants.USER_CONTEXT_KEY).(*dtos.JwtPayload)
+func (h UserContextHelpers) LoadUserFromContextWithRole(r *http.Request, role types.Role) (*models.UserModel, error) {
+	userContext, ok := r.Context().Value(service.USER_CONTEXT_KEY).(*dtos.JwtPayload)
 	if !ok {
 		return nil, ErrMissingUserContext
 	}
