+++ pkg/types/roles.go
@@ -1,0 +1,9 @@
+package types
+
+type Role string
+
+const (
+	UserRole      Role = "user"
+	AdminRole     Role = "admin"
+	OrganizerRole Role = "organizer"
+)
